Bug fixes, User Cookies

+++ KanjiOboe.Server/Interfaces/IUserRepository.cs
@@ -7,7 +7,7 @@ namespace KanjiOboe.Server.Interfaces
         Task AddUserAsync(User user);
         void DeleteUser(User user);
         Task<IEnumerable<User>> GetAllUsersAsync();
-        Task<User?> GetUserByIdAsync(int deckId);
+        Task<User?> GetUserByIdAsync(long deckId);
 
         Task<User?> GetUserByEmailAsync(string email);
         void UpdateUserAsync(User user);
